rename sync scripts now that they are universal
